fix(json): omit function/symbol members in the primitive-only template path; split the shape template out; fail the polyglot harness loudly
Three follow-ons to the #7264 array-element truncation fix.

1. Second defect in the same emitter. The template's primitive-only fast
   path is chosen by SAMPLING element 0, then trusted every later element
   to be primitive too. A function- or symbol-valued property in a later
   element must be OMITTED per SerializeJSONObject, but the fast path had
   already written the key prefix and rendered the closure via
   `stringify_value_depth` as `null` — emitting a member that must not
   exist. Only the general path pre-scanned for it. Roll the buffer back
   and defer to the slow path, exactly as the loop already did for a stray
   UNDEFINED. Applies to inline slots as much as overflow ones (an inline
   closure in a later element produced `"f2":null` too).

2. Move the shape-template machinery (`ShapeTemplate`, `shape_template_for`,
   `build_shape_prefix_template`, `try_emit_shape_element` and helpers) into
   `json/stringify_shape_template.rs`. Pure code motion — the comments the
   fix added pushed `stringify.rs` to 2062 lines, over the 2000-line lint
   gate. `stringify.rs` is now 1645, the sibling 442.

3. `benchmarks/json_polyglot/run.sh` printed "Machine-readable results:
   <path>" and exited 0 while its embedded publisher had already aborted via
   `raise SystemExit(...)` and written nothing — the script runs under
   `set -uo pipefail`, not `-e`. `run_public_baseline.sh` therefore sailed
   past it and failed ~40 minutes later at final assembly with a misleading
   "could not load json-polyglot.json".

   - A cross-runtime checksum gate now runs unconditionally right after the
     benchmark cells (not only when publishing), names the disagreeing
     runtimes, and fails the leg immediately. It mirrors the publisher's cell
     selection: perry's optimized row, node/bun's idiomatic row.
   - The publisher's exit status AND a non-empty output file are both checked
     before the success line prints.
   - A stale artifact from a previous run is removed up front so it can never
     be mistaken for this run's output.

Plus `test-files/test_gap_json_array_element_overflow_fields.ts`, byte-identical
against the pinned node 26.5.1 oracle.
